cleaned up extraneous code in the files I added
