+++ bindings/buffer.cpp
@@ -1,11 +1,6 @@
+// PyTimeloop header files.
 #include "pytimeloop/bindings/buffer.h"
 #include "pytimeloop/bindings/type_casters.h"
-
-#include <optional>
-#include <variant>
-
-// PyBind11 headers
-#include "pybind11/stl.h"
 
 // Timeloop headers
 #include "model/buffer.hpp"
