+++ include/pytimeloop/bindings/config.h
@@ -1,9 +1,7 @@
 #pragma once
 
 // PyBind11 headers
-#include "pybind11/iostream.h"
 #include "pybind11/pybind11.h"
-#include "pybind11/stl.h"
 
 #define USE_ACCELERGY
