+++ include/pytimeloop/bindings/topology.h
@@ -3,12 +3,10 @@
 // PyBind11 headers
 #include "pybind11/pybind11.h"
 
-#define USE_ACCELERGY
-
 namespace py = pybind11;
 
 namespace pytimeloop::model_bindings {
 
-void BindBufferClasses(py::module& m);
+void BindTopology(py::module& m);
 
-}
+}  // namespace pytimeloop::model_bindings
